+++ src/main/java/com/example/demo/model/repositories/ChampionshipRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.model.repositories;
 import com.example.demo.model.entities.Championship;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChampionshipRepository extends JpaRepository<Championship, Long> {
+public interface ChampionshipRepository extends JpaRepository<Championship, Integer> {
 }
